requirements.txt
streamlit>=1.31
pydantic>=2.6
python-dotenv>=1.0
pandas>=2.0
pyyaml>=6.0
openai>=1.30
